requirements.txt
Django
djangorestframework
pandas
scikit-learn
joblib
openai
python-dotenv
